[hbase] update HBase bindings for eom (#1396)
* remove 0.98, 1.0, 1.2, and 2.0 bindings
* change the 1.4 binding to be  a "HBase 1.y" binding
* add "HBase 2.y" binding and default it to the current 2.2 release
* incorporate README from 0.98 binding into current bindings
* incorporate README on bigtable testing from 1.0 binding into 1.4 binding
* incorporate implementation from 1.0 client into current bindings
* updated asynchbase binding to include parts of removed bindings it referenced
* update maprdb binding for the change in the hbase bindings
* update 1.4 and 2.2 to current releases
* use shaded client test for all hbase bindings.
* make hbase bindings consistently use log4j
* fixes #1173
* fixes #1172
