+++ bin/bindings.properties
@@ -48,11 +48,8 @@ foundationdb:site.ycsb.db.foundationdb.FoundationDBClient
 geode:site.ycsb.db.GeodeClient
 googlebigtable:site.ycsb.db.GoogleBigtableClient
 googledatastore:site.ycsb.db.GoogleDatastoreClient
-hbase098:site.ycsb.db.HBaseClient
-hbase10:site.ycsb.db.HBaseClient10
-hbase12:site.ycsb.db.hbase12.HBaseClient12
-hbase14:site.ycsb.db.hbase14.HBaseClient14
-hbase20:site.ycsb.db.hbase20.HBaseClient20
+hbase1:site.ycsb.db.hbase1.HBaseClient1
+hbase2:site.ycsb.db.hbase2.HBaseClient2
 hypertable:site.ycsb.db.HypertableClient
 ignite:site.ycsb.db.ignite.IgniteClient
 ignite-sql:site.ycsb.db.ignite.IgniteSqlClient
