+++ bin/ycsb.sh
@@ -141,6 +141,13 @@ YCSB release."
   BINDING_DIR="cassandra"
 fi
 
+# hbase14 replaced by hbas1
+if [ "${BINDING_DIR}" = "hbase14" ] ; then
+  echo "[WARN] The 'hbase14' client has been deprecated. HBase 1.y users should \
+rely on the 'hbase1' client instead."
+  BINDING_DIR="hbase1"
+fi
+
 # arangodb3 deprecation message
 if [ "${BINDING_DIR}" = "arangodb3" ] ; then
   echo "[WARN] The 'arangodb3' client has been deprecated. The binding 'arangodb' \
@@ -230,20 +237,6 @@ fi
 if [ "${BINDING_DIR}" = "couchbase" ] ; then
   echo "[WARN] The 'couchbase' client is deprecated. If you are using \
 Couchbase 4.0+ try using the 'couchbase2' client instead."
-fi
-
-# HBase 0.98 deprecation message
-if [ "${BINDING_DIR}" = "hbase098" ] ; then
-  echo "[WARN] The 'hbase098' client is deprecated because HBase 0.98 \
-is EOM. If you are using HBase 1.2+ try using the 'hbase12' client \
-instead."
-fi
-
-# HBase 1.0 deprecation message
-if [ "${BINDING_DIR}" = "hbase10" ] ; then
-  echo "[WARN] The 'hbase10' client is deprecated because HBase 1.0 \
-is EOM. If you are using HBase 1.2+ try using the 'hbase12' client \
-instead."
 fi
 
 # For Cygwin, switch paths to Windows format before running java
