--- hbase098/src/main/java/site/ycsb/db/package-info.java
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2017, Yahoo!, Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License. See accompanying LICENSE file.
- */
-
-/**
- * The YCSB binding for <a href=
- * "https://issues.apache.org/jira/browse/HBASE/fixforversion/12333364/">HBase
- * 0.98.X</a>.
- */
-package site.ycsb.db;
