+++ hbase1/src/test/resources/hbase-site.xml
