+++ hbase1/src/test/resources/log4j.properties
