--- hbase12/pom.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright (c) 2012 - 2016 YCSB contributors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License"); you
-may not use this file except in compliance with the License. You
-may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-implied. See the License for the specific language governing
-permissions and limitations under the License. See accompanying
-LICENSE file.
--->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>site.ycsb</groupId>
-    <artifactId>binding-parent</artifactId>
-    <version>0.18.0-SNAPSHOT</version>
-    <relativePath>../binding-parent/</relativePath>
-  </parent>
-
-  <artifactId>hbase12-binding</artifactId>
-  <name>HBase 1.2 DB Binding</name>
-
-  <properties>
-    <skipJDK9Tests>true</skipJDK9Tests>
-    <skipJDK10Tests>true</skipJDK10Tests>
-    <skipJDK11Tests>true</skipJDK11Tests>
-
-    <!-- Tests can't run without a shaded hbase testing util.
-         See HBASE-15666, which blocks us.
-         For now, we rely on the HBase 1.0 binding and manual testing.
-      -->
-    <maven.test.skip>true</maven.test.skip>
-  </properties>
-  <dependencies>
-    <dependency>
-      <groupId>site.ycsb</groupId>
-      <artifactId>hbase10-binding</artifactId>
-      <version>${project.version}</version>
-      <!-- Should match all compile scoped dependencies -->
-      <exclusions>
-        <exclusion>
-          <groupId>org.apache.hbase</groupId>
-          <artifactId>hbase-client</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-    <dependency>
-      <groupId>site.ycsb</groupId>
-      <artifactId>core</artifactId>
-      <version>${project.version}</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.hbase</groupId>
-      <artifactId>hbase-shaded-client</artifactId>
-      <version>${hbase12.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.12</version>
-      <scope>test</scope>
-    </dependency>
-<!-- blocked on HBASE-15666
-    <dependency>
-      <groupId>org.apache.hbase</groupId>
-      <artifactId>hbase-testing-util</artifactId>
-      <version>${hbase12.version}</version>
-      <scope>test</scope>
-      <exclusions>
-        <exclusion>
-          <groupId>jdk.tools</groupId>
-          <artifactId>jdk.tools</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
--->
-  </dependencies>
-</project>
