+++ hbase2/src/test/resources/log4j.properties
